Move the map iterator
